Recent changes within AROS made it necessary to rename the library functions and add a parameter to the SLIB macros.
That additional number is the LVO. We have to set it to 0 because
of the way how the compatibility macros are defined, e.g.:

#define LIBFUNC_P2(returntype, funcname,\
register1, type1, name1,\
register2, type2, name2)\
AROS_LH1(returntype, funcname,\
AROS_LHA(type1, name1, register1),\
type2, name2, 0, name2)\
             ^^^
{\
AROS_LIBFUNC_INIT

That's OK because it's only a text replacement. The real LVOs are
in include/defines.